moved nte examples and lbc examples to separate subdirectories
