+++ lbc/problems/nte_examples/example3.py
