+++ lbc/problems/nte_examples/example6.py
