+++ lbc/problems/nte_examples/example7.py
